+++ src/main/java/org/apiphany/security/client/SecuredExchangeClientBuilder.java
@@ -2,7 +2,9 @@ package org.apiphany.security.client;
 
 import java.util.function.Consumer;
 
+import org.apiphany.client.ExchangeClient;
 import org.apiphany.client.ExchangeClientBuilder;
+import org.apiphany.lang.ScopedResource;
 import org.apiphany.security.oauth2.client.OAuth2HttpExchangeClientBuilder;
 import org.morphix.lang.function.Consumers;
 
@@ -12,6 +14,11 @@ import org.morphix.lang.function.Consumers;
  * @author Radu Sebastian LAZIN
  */
 public class SecuredExchangeClientBuilder extends ExchangeClientBuilder {
+
+	/**
+	 * Indicates whether a security mechanism was defined.
+	 */
+	private boolean securityDefined = false;
 
 	/**
 	 * Hide constructor.
@@ -30,13 +37,37 @@ public class SecuredExchangeClientBuilder extends ExchangeClientBuilder {
 	}
 
 	/**
+	 * This builder always requires a security mechanism to be defined.
+	 *
+	 * @return never returns normally
+	 * @throws IllegalStateException always thrown to indicate no security mechanism was defined
+	 */
+	@Override
+	public ScopedResource<ExchangeClient> build() {
+		if (securityDefined) {
+			return super.build();
+		}
+		throw new IllegalStateException("Client not secured with any mechanism");
+	}
+
+	/**
+	 * Marks that a security mechanism was defined.
+	 *
+	 * @return this builder
+	 */
+	private SecuredExchangeClientBuilder securityDefined() {
+		this.securityDefined = true;
+		return this;
+	}
+
+	/**
 	 * Adds OAuth2 functionality.
 	 *
 	 * @param oAuth2BuilderCustomizer OAuth2 customizer
 	 * @return new OAuth2 exchange client builder
 	 */
 	public OAuth2HttpExchangeClientBuilder oAuth2(final Consumer<OAuth2HttpExchangeClientBuilder> oAuth2BuilderCustomizer) {
-		return decoratedWithBuilder(OAuth2HttpExchangeClientBuilder.class, oAuth2BuilderCustomizer);
+		return securityDefined().decoratedWithBuilder(OAuth2HttpExchangeClientBuilder.class, oAuth2BuilderCustomizer);
 	}
 
 	/**